+++ HireHub/Controllers/CompanyController.cs
@@ -29,11 +29,26 @@ namespace HireHub.Web.Controllers
             {
                 return RedirectToAction("Edit", "Company");
             }
-            return View();
+            else
+            {
+                if (User.IsInRole("Admin"))
+                {
+                    TempData["ErrorMessage"] = "Admin can't create a company!";
+                    return RedirectToAction("Index", "Home");
+                }
+                TempData["ErrorMessage"] = "You must create a company first!";
+                return View();
+            }
         }
         [HttpGet]
         public async Task<IActionResult> Edit()
         {
+            var isExist = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isExist)
+            {
+                TempData["ErrorMessage"] = "You need to have a company";
+                return RedirectToAction("Create", "Company");
+            }
             var model = await _companyService.GetCompanyByUserId(GetUserId());
 
             return View(model);
@@ -42,6 +57,12 @@ namespace HireHub.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(EditCompanyVM model)
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You need to have a company to edit!";
+                return RedirectToAction("Create", "Company");
+            }
             if (!ModelState.IsValid)
             {
                 TempData["ErrorMessage"] = "Enter invalid copmany data!";
@@ -62,6 +83,12 @@ namespace HireHub.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(CreateCompanyVM createCompanyVM)
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You alreay have a company";
+                return RedirectToAction("Explore", "Job");
+            }
             if (!ModelState.IsValid)
             {
                 return View(createCompanyVM);
@@ -83,6 +110,12 @@ namespace HireHub.Web.Controllers
         [HttpGet]
         public async Task<IActionResult> MyApplication()
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You need to have a company!";
+                return RedirectToAction("Create", "Company");
+            }
             int companyId = await _companyService.GetCompanyIdByUserId(GetUserId());
 
             var myApplication = await _companyService.MyApplication(companyId);
@@ -93,6 +126,12 @@ namespace HireHub.Web.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You need to have a company!";
+                return RedirectToAction("Create", "Company");
+            }
             try
             {
                 await _companyService.DeleteCompany(id);
@@ -108,6 +147,12 @@ namespace HireHub.Web.Controllers
 
         public async Task<IActionResult> Hire(string id,string email)
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You need to have a company!";
+                return RedirectToAction("Create", "Company");
+            }
             var getJob = await _jobService.GetJobDetails(id);
             string subject = $"Congratulations! You've been hired as a {getJob!.Title}";
             string message = $"Dear {email},<br/>" +
@@ -150,6 +195,12 @@ namespace HireHub.Web.Controllers
 
         public async Task<IActionResult> Reject(string id , string email)
         {
+            var isHaveCompany = await _companyService.IsUserHaveCompany(GetUserId());
+            if (!isHaveCompany)
+            {
+                TempData["ErrorMessage"] = "You need to have a company!";
+                return RedirectToAction("Create", "Company");
+            }
             var userId = await _companyService.GetUserIdByEmail(email);
 
             bool? isHiring = await _companyService.IsHire(userId, id);
